\name{bsmim2-package}
\alias{bsmim2-package}
\alias{bsmim2}
\docType{package}
\title{
\packageTitle{bsmim2}
}
\description{
\packageDescription{bsmim2}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{bsmim2}
\packageIndices{bsmim2}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{bsmim2}

Maintainer: \packageMaintainer{bsmim2}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
