% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsmim2.R
\name{bsmim2}
\alias{bsmim2}
\title{Bayesian Structural Multipollutant Index Model (BSMIM)}
\usage{
bsmim2(
  y,
  x,
  z,
  group_id = NULL,
  niter,
  nburn = round(niter/2),
  nthin = 1,
  prior_sigma = c(0.001, 0.001),
  prior_lambda = 1,
  prior_lambdaB = 1,
  prior_lambda_shaperate = c(1, 0.1),
  prior_tau = 0.01,
  kappa = 1,
  basis.opts.list = NULL,
  horseshoe = FALSE,
  spike_slab = FALSE,
  gauss_prior = FALSE,
  prior_theta_slab_sd = 0.25,
  prior_theta_slab_bounds = c(0, 100),
  prior_pi = c(1, 1),
  stepsize_theta = 0.1,
  gaussian = TRUE,
  polydegree = 2,
  draw_h = FALSE,
  constraints = NULL,
  prior_slabpos = c(0.4, 1.6),
  prior_alphas = NULL,
  prior_slabrho = c(4, 2)
)
}
\arguments{
\item{y}{Vector of outcomes.}

\item{x}{A list of M matrices. Each element is an (N by L_m) matrix, with each row being an individual's L_m-vector of exposures in the mth index.}

\item{z}{A matrix of covariates and confounders. This can be ommited.}

\item{group_id}{A vector of ids indicating cluster membership for random intercepts model.}

\item{niter}{Number of MCMC iterations including burnin.}

\item{nburn}{The number of iterations to be discarded as burnin. Must be less than niter.}

\item{nthin}{Thinning, every nthin-th draw from the posterior will be saved.}

\item{prior_sigma}{Vector of length 2 corresponding to hyperparameters \eqn{(a_{\sigma},b_{\sigma})}, the parameters for the gamma prior on \eqn{\sigma^{-2}}}

\item{prior_lambda}{Hyperparameter \eqn{b_{\lambda}}, the prior variance on \eqn{log(\lambda^{-1})}, (only used in the horseshoe/non-spike-slab model.}

\item{prior_lambdaB}{Hyperparameter \eqn{b_{\lambda_B}}, the prior variance on \eqn{log(\lambda^{-1}_B)},.}

\item{prior_lambda_shaperate}{Hyperparameters (\eqn{a_{\lambda}}, \eqn{b_{\lambda}}), the shape and rate parameters for gamma prior.}

\item{prior_tau}{Hyperparameter \eqn{\tau_{0}^2}, scale parameter for half Cauchy prior for \eqn{\tau}. (default 1)}

\item{kappa}{Vector of length M of scale parameters, kappa_m. If only one number is provided, we take all \eqn{\kappa_m}=kappa}

\item{basis.opts.list}{List with the entries: type = the type of basis used, either 'face' (default) or "ns" or "bs" for splines or "gam" for presmoothing the exposure with a gam following defaults from mgcv; knots = the number of knots used for method face; pve = the percent of variance explained by the PCs for method face; df = the df for ns method.}

\item{horseshoe}{Use the horseshoe prior for indexwise selection (default FALSE)}

\item{gaussian}{Use a Gaussian kernel (TRUE, default) or a polynomial kernel (FALSE)}

\item{polydegree}{Degree of polynomial when polynomial kernel is used.  Only applies when gaussian=FALSE.}
}
\value{
An object of class 'bsmim'.
}
\description{
This function fits the BSMIM. Only gaussian kernel is implemented as of now. Adapted from "bkmrdlm_multi.cpp" in the "regimes" package by Ander Wilson.
}
\author{
Glen McGee and Ander Wilson (adapted from the "regimes" package by Ander Wilson).
}
