% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsmim2.R
\name{bsmim_crossval2}
\alias{bsmim_crossval2}
\title{K-Fold Cross Validation for BSMIM}
\usage{
bsmim_crossval2(object, kfolds = 4, correct = FALSE)
}
\arguments{
\item{object}{an object of class 'bsmim'}
}
\value{
An list of class 'bsmim'.
}
\description{
K-Fold Cross Validation for BSMIM
}
\author{
Glen McGee and Ander Wilson (adapted from the "regimes" package by Ander Wilson).
}
