% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bsmim_predict_approx_cpp2}
\alias{bsmim_predict_approx_cpp2}
\title{Predicting hnew for BSMIM by Component using APPROXIMATE method, computing the conditional mean and variance of hnew given posterior means of everything else}
\usage{
bsmim_predict_approx_cpp2(
  yz,
  Xlist,
  thetalist,
  psilist,
  rho,
  gamma,
  lambdaInverse,
  lambdaBInverse,
  sigma2,
  weightslist,
  gridpointslist,
  poly,
  d,
  randint,
  Bmat
)
}
\arguments{
\item{yz}{matrix [Y,Z], Z does not include a vector of 1s (i.e. no intercept)}

\item{Xlist}{list of (N by L_m) matrices representing X_m}

\item{thetalist}{list of (N by L_m) matrices representing X_m}

\item{psilist}{list of basis matrices}

\item{rho}{(S by M) matrix of rho_m draws}

\item{gamma}{list of (S by P_z) matrices with gamma draws}

\item{lambdaInverse}{S-vector of lambda^{-1} draws}

\item{lambdaBInverse}{S-vector of lambda^{-1}_B draws}

\item{sigma2}{S-vector of sigma^2 draws}

\item{weightslist}{list of (S by L_m) weight matrices to apply to exposures X_m (basically A theta^*, so that X*theta*=x(Atheta*))}

\item{gridpointslist}{List of (points by L_m) matrix of new values to predict at}

\item{poly}{0 = gaussian kernel, 1 = polynomial kernel}

\item{d}{degree of polynomial kernel}

\item{randint}{0 = no random intercepts, 1 = random intercepts model}

\item{Bmat}{N xN block diagonal matrix indicating cluster membership for random intercepts model}
}
\description{
Returns a list containing (points by (sum_m Lm) ) matrices of hmean and hvar
}
\author{
Glen McGee (adapted from the "regimes" package by Ander Wilson)
}
