% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bsmim_predict_indexwise_cpp2}
\alias{bsmim_predict_indexwise_cpp2}
\title{Predicting hnew for BSMIM by Index}
\usage{
bsmim_predict_indexwise_cpp2(
  yz,
  Xlist,
  thetalist,
  psilist,
  rho,
  gamma,
  lambdaInverse,
  lambdaBInverse,
  sigma2,
  weightslist,
  gridpoints,
  Xmedianlist,
  Eq,
  crossM,
  poly,
  d,
  randint,
  Bmat
)
}
\arguments{
\item{yz}{matrix [Y,Z], Z does not include a vector of 1s (i.e. no intercept)}

\item{Xlist}{list of (N by L_m) matrices representing X_m}

\item{thetalist}{list of (N by L_m) matrices representing X_m}

\item{psilist}{list of basis matrices}

\item{rho}{(S by M) matrix of rho_m draws}

\item{gamma}{list of (S by P_z) matrices with gamma draws}

\item{lambdaInverse}{S-vector of lambda^{-1} draws}

\item{lambdaBInverse}{S-vector of lambda^{-1}_B draws}

\item{sigma2}{S-vector of sigma^2 draws}

\item{weightslist}{list of (S by L_m) weight matrices to apply to exposures X_m (basically A theta^*, so that X*theta*=x(Atheta*))}

\item{gridpoints}{(points by M) matrix containing grid of new index levels for each index m}

\item{Xmedianlist}{list of L_m-vectors containing exposure medians}

\item{Eq}{M-vector of index quantiles}

\item{crossM}{index to set cross section}

\item{poly}{0 = gaussian kernel, 1 = polynomial kernel}

\item{d}{degree of polynomial kernel}

\item{randint}{0 = no random intercepts, 1 = random intercepts model}

\item{Bmat}{N xN block diagonal matrix indicating cluster membership for random intercepts model}
}
\description{
Predicting hnew for BSMIM by Index
}
\author{
Glen McGee and Ander Wilson (adapted from the "regimes" package by Ander Wilson)
}
