% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bsmim_spikeslab_gaussprior_mcmc2}
\alias{bsmim_spikeslab_gaussprior_mcmc2}
\title{MCMC for BSMIM with spike and slab prior (inv-uniform slab for rho version)}
\usage{
bsmim_spikeslab_gaussprior_mcmc2(
  yz,
  Xlist,
  a_lam,
  b_lam,
  b_lambdaB,
  a_sig,
  b_sig,
  s_theta,
  step_theta,
  a_pi,
  b_pi,
  poly,
  d,
  randint,
  Bmat,
  draw_h,
  n_inner,
  n_outer,
  n_burn
)
}
\arguments{
\item{yz}{matrix [Y,Z], Z does not include a vector of 1s (i.e. no intercept)}

\item{Xlist}{list of (N by L_m) matrices representing X_m}

\item{a_lam}{shape hyperparameter for \eqn{\lambda^{-1}}}

\item{b_lam}{rate hyperparameter for \eqn{\lambda^{-1}}}

\item{b_lambdaB}{hyperparameter for \eqn{\lambda^{-1}_B}}

\item{a_sig}{first hyperparameter for \eqn{\sigma^{-2}}}

\item{b_sig}{second hyperparameter for \eqn{\sigma^{-2}}}

\item{poly}{0 = gaussian kernel, 1 = polynomial kernel}

\item{d}{degree of polynomial kernel}

\item{randint}{0 = no random intercepts, 1 = random intercepts model}

\item{Bmat}{N xN block diagonal matrix indicating cluster membership for random intercepts model}

\item{draw_h}{0 = dont draw h, 1 = draw h}

\item{n_inner}{no. of MCMC iterations to run in the inner loop. n_outer*n_inner iteraction will be run.}

\item{n_outer}{no. of MCMC iterations to run in the outer loop. n_outer iterations will be saved.}

\item{n_burn}{no. of MCMC iterations to discard as burn-in}

\item{tau02}{hyperparameter for tau2; classic horseshoe is 1}

\item{kappa}{vector of hyperparameters \eqn{\kappa_m} for \eqn{\theta^*_m}}

\item{horseshoe}{0 = no selection, 1 = componentwise horseshoe priors}
}
\description{
Returns posterior draws for \eqn{\theta},\eqn{\lambda^{-1}},\eqn{\rho},\eqn{\sigma^2},\eqn{\gamma}, as well as the mean(\eqn{h}) and cov(\eqn{h})
Adapted from "bkmrdlm_multi.cpp" in the "regimes" package by Ander Wilson.
Note that this implementation with horseshoe2 prior returns posterior draws of \eqn{\rho} as a list of M matrices (unlike main MCMC code, which returns one matrix of M columns)
}
\author{
Glen McGee and Ander Wilson (adapted from the "regimes" package by Ander Wilson).
}
