% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsmim2.R
\name{check_conv}
\alias{check_conv}
\title{Summarize convergence via PSR factors
#'}
\usage{
check_conv(fit)
}
\arguments{
\item{obj}{An object of class bsmim}
}
\value{
A list PSR factors
}
\description{
Summarize convergence via PSR factors
#'
}
\author{
Glen McGee
}
