% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsmim2.R
\name{get_PSR}
\alias{get_PSR}
\title{Compute PSR 
#'}
\usage{
get_PSR(x, nchains)
}
\arguments{
\item{x}{a vector of pooled mcmc draws}

\item{nchains}{number of mcmc chains}
}
\value{
A list PSR factors
}
\description{
Compute PSR 
#'
}
\author{
Glen McGee
}
