% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trace2.R
\name{plot_trace2}
\alias{plot_trace2}
\title{Plot component-wise response curves}
\usage{
plot_trace2(object)
}
\arguments{
\item{object}{An object of class bsmim}
}
\value{
a list of trace plots.
}
\description{
Plot component-wise response curves
}
