% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_univar_hnew2.R
\name{plot_univar_hnew2}
\alias{plot_univar_hnew2}
\title{Plot component-wise response curves}
\usage{
plot_univar_hnew2(pred, ylims = NULL, assoc = TRUE, centered = TRUE)
}
\arguments{
\item{pred}{An object of class hpred}

\item{ylims}{An optional vector of length 2 containing the limits for the y axis; default is null}

\item{assoc}{Report contrasts}

\item{centered}{Remove intercept; default is TRUE}
}
\value{
a list of component-wise response-curve plots .
}
\description{
Plot component-wise response curves
}
