% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_univar_hnew2.R
\name{plot_univar_hnew_indexwise2}
\alias{plot_univar_hnew_indexwise2}
\title{#' Old version: Plot component-wise response curves (only difference is the format of pred$grid)
#' 
#'
#' @param pred An object of class hpred
#' @param ylims An optional vector of length 2 containing the limits for the y axis; default is null
#'
#' @return a list of component-wise response-curve plots .
#' @export
#'
plot_univar_hnew_old2 <- function(pred,ylims=NULL){
  
  ## make dataframe from object of class hpred
  df <- data.frame(matrix(unlist(pred), nrow=length(pred$m), byrow=F))
  colnames(df) <- names(pred)
  
  ## set ylimits
  if(length(ylims)!=2){
    miny <- min(df$lower);maxy <- max(df$upper)  
  }else if(ylims[1]>=ylims[2]){
    miny <- min(df$lower);maxy <- max(df$upper)  
  }else{
    miny <- ylims[1];maxy <- ylims[2]
  }
  
  plots <- list()
  for(mm in 1:max(df$m)){                 ## loop over m (indices)
    plots_m <- list()
    for(ll in 1:max(df$l[df$m==mm])){     ## loop over l (components)
      
      dfplot <- df[df$m==mm & df$l==ll,]  ## subset data
      
      p <- ggplot(dfplot, aes(x=grid, y=mean))+
        geom_line(linetype=3)+ ## 3 is dotted
        geom_ribbon(aes(ymin=lower,ymax=upper),alpha=0.2)+
        ylim(miny,maxy)+
        ylab("Estimated exposure-response (h)")+
        xlab("Exposure Component")+
        ggtitle(paste("Index",mm,", Component",ll))+
        theme_bw() +
        theme(panel.grid.major = element_blank(),panel.grid.minor = element_blank()) 
      plots_m[[ll]] <- p
    }
    plots[[mm]] <- plots_m
  }
  
  return(plots)
}
Plot component-wise response curves}
\usage{
plot_univar_hnew_indexwise2(pred, ylims = NULL)
}
\arguments{
\item{pred}{An object of class hpred_indexwise}

\item{ylims}{An optional vector of length 2 containing the limits for the y axis; default is null}
}
\value{
a list of component-wise response-curve plots .
}
\description{
#' Old version: Plot component-wise response curves (only difference is the format of pred$grid)
#' 
#'
#' @param pred An object of class hpred
#' @param ylims An optional vector of length 2 containing the limits for the y axis; default is null
#'
#' @return a list of component-wise response-curve plots .
#' @export
#'
plot_univar_hnew_old2 <- function(pred,ylims=NULL){
  
  ## make dataframe from object of class hpred
  df <- data.frame(matrix(unlist(pred), nrow=length(pred$m), byrow=F))
  colnames(df) <- names(pred)
  
  ## set ylimits
  if(length(ylims)!=2){
    miny <- min(df$lower);maxy <- max(df$upper)  
  }else if(ylims[1]>=ylims[2]){
    miny <- min(df$lower);maxy <- max(df$upper)  
  }else{
    miny <- ylims[1];maxy <- ylims[2]
  }
  
  plots <- list()
  for(mm in 1:max(df$m)){                 ## loop over m (indices)
    plots_m <- list()
    for(ll in 1:max(df$l[df$m==mm])){     ## loop over l (components)
      
      dfplot <- df[df$m==mm & df$l==ll,]  ## subset data
      
      p <- ggplot(dfplot, aes(x=grid, y=mean))+
        geom_line(linetype=3)+ ## 3 is dotted
        geom_ribbon(aes(ymin=lower,ymax=upper),alpha=0.2)+
        ylim(miny,maxy)+
        ylab("Estimated exposure-response (h)")+
        xlab("Exposure Component")+
        ggtitle(paste("Index",mm,", Component",ll))+
        theme_bw() +
        theme(panel.grid.major = element_blank(),panel.grid.minor = element_blank()) 
      plots_m[[ll]] <- p
    }
    plots[[mm]] <- plots_m
  }
  
  return(plots)
}
Plot component-wise response curves
}
