% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_weights_bsmim2.R
\name{plot_weights_bsmim2}
\alias{plot_weights_bsmim2}
\title{Plot weights (rho, theta, theta*) for bsmim}
\usage{
plot_weights_bsmim2(object)
}
\arguments{
\item{object}{An object of class bsmim}
}
\value{
a list of boxplots.
}
\description{
Plot weights (rho, theta, theta*) for bsmim
}
