% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsmim_predict2.R
\name{predict_hnew_X2}
\alias{predict_hnew_X2}
\title{Predict hnew for BSMIM for new exposure values}
\usage{
predict_hnew_X2(
  object,
  newX = NULL,
  newY = NULL,
  newZ = NULL,
  newAmat = NULL,
  rawX = TRUE,
  qtl_lims = c(0.05, 0.95),
  points = 20
)
}
\arguments{
\item{object}{An object of class bsmim}

\item{newX}{M-list of (pts by Lm)-matrices containing new values to predict at---all others get set to 0; overrides qtl and points}

\item{newY}{N-vector of new outcomes}

\item{newZ}{N by P_z matrix of new covariates}

\item{newAmat}{N by K design matrix indicating cluster membership (only used when using random intercepts)}
}
\value{
a list of (1) a dataframe containing mean, sd and interval for hnew, (2) matrix of samples of hnew, (3) list of exposure grid (newX)
}
\description{
Predict hnew for BSMIM for new exposure values
}
