% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsmim_predict2.R
\name{predict_hnew2}
\alias{predict_hnew2}
\title{Predict hnew for BSMIM}
\usage{
predict_hnew2(
  object,
  points = 20,
  qtl_lims = c(0.05, 0.95),
  qtls = list(NA),
  newX = NULL,
  approx = FALSE
)
}
\arguments{
\item{object}{An object of class bsmim}

\item{points}{Number of points to predict at}

\item{qtl_lims}{Quantile limits for grid of exposure values}

\item{qtls}{M-list of Lm-vectors containing quantiles to set exposure components to}

\item{newX}{optional M-list of (pts by Lm)-matrices containing new values to predict at---all others get set to 0; overrides qtls and points}

\item{approx}{use approximate method (not recommended)}
}
\value{
a dataframe containing predicted values.
}
\description{
Predict hnew for BSMIM
}
