% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsmim_predict2.R
\name{predict_hnew_indexwise2}
\alias{predict_hnew_indexwise2}
\title{Predict hnew for BSMIM}
\usage{
predict_hnew_indexwise2(
  object,
  points = 20,
  qtl_lims = c(0.05, 0.95),
  crossM = 0,
  qtl = 0.5,
  grid_lims = c(-2, 2),
  abs_grid = FALSE,
  trueW = NULL
)
}
\arguments{
\item{object}{An object of class bsmim}

\item{points}{Number of points to predict at}

\item{qtl_lims}{Quantile limits for grid of exposure values}

\item{crossM}{Exposure to set a cross section quantile for}

\item{qtl}{quantile for the cross section}

\item{grid_lims}{Optional (absolute) index limits for grid of index values (alternative to the default use of quantiles)}

\item{abs_grid}{Should the absolute index limits be used (TRUE), or the standard quantile limits (FALSE; default)}

\item{trueW}{weights for prediction for comparison to true curve; if NULL then set to posterior mean}
}
\value{
a dataframe containing predicted values.
}
\description{
Predict hnew for BSMIM
}
