% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsmim_predict2.R
\name{predict_hnew_old2}
\alias{predict_hnew_old2}
\title{Predict hnew for BSMIM (old version)}
\usage{
predict_hnew_old2(
  object,
  points = 20,
  qtls = list(NA),
  newvals = NULL,
  approx = FALSE
)
}
\arguments{
\item{object}{An object of class bsmim}

\item{points}{Number of points to predict at}

\item{qtls}{M-list of Lm-vectors containing quantiles to set exposure components to}

\item{newvals}{optional M-list of (pts by Lm)-matrices containing new values to predict at---all others get set to 0; overrides qtls and points}

\item{approx}{should the approximate method be used (not recommended)}
}
\value{
a dataframe containing predicted values.
}
\description{
Predict hnew for BSMIM (old version)
}
