% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsmim2.R
\name{summarize_thetas}
\alias{summarize_thetas}
\title{Summarize posterior of thetas (weights)}
\usage{
summarize_thetas(obj, w = FALSE)
}
\arguments{
\item{y}{obj An object of class bsmim}
}
\value{
A list of length M of data.frames summarizing theta estimates.
}
\description{
This function takes as input a bsmim object and summarizes the distribution of estimates of the weights, theta.
}
\author{
Glen McGee
}
